requirements.txt
streamlit
langchain_openai
pypdf
python-dotenv
tiktoken